refactor: HttpServer to Server
Changes:
    - Rename class from HttpServer to Server
    - Remove TypeGraphQL
    - Change from express-graphql to apollo-server
    - Remove TypeOrm data source initialization
